requirements.txt
python-telegram-bot==20.8
openai==1.30.1
python-dotenv==1.0.1